Removed duplicate GMSH classes
